fix: delivery, tenancy and GDPR-erasure correctness across the outbox, CDC, transports and authorization
Accumulated work since 10.0.0-alpha.11. The theme is silent wrongness: defects
where the framework reported success while losing, duplicating, or failing to
erase data, and nothing downstream could tell.

Compliance and GDPR erasure
- An erasure contributor now discharges the locations it actually erased, so an
  erasure request can reach Completed instead of s
- Certificates stored on PostgreSQL verify instead of reporting TAMPERED to the
  consumer; SQL stores round-trip every signed cla
- Key destruction can complete on Azure Key Vault and AWS KMS.
- Completed is no longer reported while encrypted
  readable under a surviving key.
- Registrations now declare their store kind. Exis
  Unknown, which discharges nothing, so erasure refuses until they are
  classified. See the migration note for the one-s

Messaging and delivery
- Kafka: a poison record is dead-lettered or tombstoned rather than stalling the
  partition forever; a rebalance no longer commits
  handlers; commit failure is no longer treated as handler failure.
- Outbox: a failure report for a terminal message r
  than ClaimLost; dead fencing branches removed from MongoDB and Redis.
- Inbox: a zero retry bound no longer makes failur
  happen; failures return Undecided instead of destroying the caller's exception.
- Saga: a timeout whose dispatch failed is no long
  deleted; dedup keys no longer collide on simple type names.
- IBM MQ no longer reports settlement the broker r

Change data capture
- SQL Server no longer drops captured changes on a second poll, and cannot write
  a durable checkpoint on a zero-row fetch.
- Overlapping PostgreSQL invocations are refused; checkpoint comparison is
  numeric rather than lexical.
- Opt-in bound on consecutive transient failures; exhaustion stops without
  writing a position.

Authorization and tenancy
- Grant and activity-group synchronisation replace atomically per user and per
  tenant instead of deleting the estate and repopu
- Grant queries filter by value rather than by empty string, so access reviews
  and entitlement reports return the grants that e
- ITenantContext resolves by precedence rather than by registration order.

+++ tests/unit/Excalibur.Compliance.Tests/Soc2/Validators/Soc2EffectivenessScoreShould.cs
@@ -46,8 +46,8 @@ public sealed class Soc2EffectivenessScoreShould
 			DateTimeOffset.UtcNow.AddDays(-1), DateTimeOffset.UtcNow));
 		var unverified = await Sut.ValidateAsync("SEC-004", CancellationToken.None);
 
-		tampered.EffectivenessScore.ShouldBeLessThan(
-			unverified.EffectivenessScore,
+		((int)tampered.EffectivenessScore).ShouldBeLessThan(
+			(int)unverified.EffectivenessScore,
 			"a violation we found is knowledge; a window we never exercised is not, and the proven "
 			+ "deficiency must never read as the better of the two.");
 	}
@@ -63,7 +63,8 @@ public sealed class Soc2EffectivenessScoreShould
 		IntegrityCheckReports(Tampered());
 		var tampered = await Sut.ValidateAsync("SEC-004", CancellationToken.None);
 
-		nothingConfigured.EffectivenessScore.ShouldBeLessThanOrEqualTo(tampered.EffectivenessScore);
+		((int)nothingConfigured.EffectivenessScore)
+			.ShouldBeLessThanOrEqualTo((int)tampered.EffectivenessScore);
 	}
 
 	[Fact]
